Started to distinguish between elements like <title>, which always have to have some human-readable content, and those like <premiere>, which may have an 'explanation' inside but do not require it.  The former have handler 'with-lang', the latter now 'with-lang-or-empty' which can return [] for an empty element.  Trying to write out inappropriate empty content (like <title></title>) will now give a warning.
Updated tv_grep to cope with this: it is still necessary to give a regular
expression as the argument to predicates like --premiere, but a regexp of
'' will match all <premiere> elements, even empty ones.  So the old
behaviour is preserved.  Added some new test cases for this (not
complete coverage though) and an example of <last-chance> to test.xml.