Now that the filter programmes can take more than one file on the command line, we can write some tests that use this.  There are a few for the sake of testing the multifile functionality, but also we can speed up the test suite because what used to take ten runs of tv_foo can now be done in one run giving ten input files.
We don't test all possible lists of files, only a few of the possible
pairs, but still it bloats the test suite to unmanageably large
numbers of tests (over 1000).  I've remedied that by defining a
smaller set of tests, including running each filter program on every
input file at once, and making this the default.  If you have a fast
computer or time to spare you can get the complete set of tests with
the --full option.

The files like tv_cat_all_UTF8.expected are for the testing of every
input file (actually, every input file with UTF-8 encoding) at once.