get_nice: support compressed response bodies
Some HTTP servers, like yle.fi, randomly return a gzip-compressed body
in the response. Switch to $r->decoded_content(charset => "none") to
automatically de-compress the body, but leave character set processing
to the caller.

This has to be implemented in XMLTV::Get_nice or otherwise this won't
work together with Memoize.